Ship the kernel as ONE bundled npm package (decision 12 publish revision)
The owner decided to publish, and that four separate scoped packages made no
sense for a single-gateway product — the 4-way split was a conversion artifact
(1:1 oracle mapping) and the real import graph has one leaf. So the kernel now
ships as ONE package: @panaversity/ksor-content-gateway bundles platform +
content + gateway-kit (workspace devDeps, tsdown noExternal) and exposes BOTH
bins — ksor-content (corpus setup: schema/ingest/calibrate) and
ksor-content-gateway (serve). An adopter installs one thing for the served rung.

- content's CLI split so its logic is a library (commands.ts → runContentCli)
  the bundled package reuses as a second bin; cli.ts stays the thin bin the
  boundary suite requires nothing to import.
- content-gateway: noExternal the @panaversity/* workspace packages; declare
  the now-direct external deps (@google/genai, jose) alongside the rest; two
  bins; files ["dist","schema"]. schema.sql is build-copied into
  content-gateway/schema/ (gitignored, shipped, resolved via import.meta.url).
- platform/content/gateway-kit revert to private-only (dev/test), never
  published; guard rule 5's content-gateway allowlist updated accordingly.

Verified with a real `pnpm pack` → `npm install` in a fresh directory: both
bins run, `ksor-content schema --dim 8` renders 316 lines (schema resolves),
deps are self-contained (catalog:/workspace: rewritten by pnpm). Full gate
green incl. the gateway DB acceptance driving the BUNDLED bin end-to-end.

Re-extract platform/gateway-kit as shared packages only when a second gateway
is a real consumer. content-gateway stays private:true until the serve wiring
lands (next); the actual publish is owner-triggered and needs the one-time npm
bootstrap (new scoped packages can't first-publish via OIDC).
